pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "koi-demo"
version = "0.1.0"
description = "KOI-net Self-Forming Knowledge Mesh Demo"
authors = [
    {name = "KOI-net Team"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "pydantic-settings",
    "httpx",
    "python-dotenv",
    "rid-lib>=3.2.1",
    "koi-net==1.0.0b12",
    "ruamel.yaml"
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools]
package-dir = {"" = "."}

# Add setuptools configuration to control package discovery
[tool.setuptools.packages.find]
where = ["."]  # Search in the root directory
include = ["nodes*"]  # Include only 'nodes' and its sub-packages
exclude = ["config*"] # Exclude the 'config' directory
namespaces = false # Assume 'nodes' itself is not a namespace package

